Organized and worked on deal inside gamestate
