--- Assignment1/src/main/java/controllers/GameState.java
@@ -1,28 +1,0 @@
-package controllers;
-import models.Deck;
-import sun.plugin.javascript.navig.Array;
-
-import java.util.ArrayList;
-
-/**
- * Created by isaac on 1/21/2016.
- */
-public class GameState {
-    Deck deck[];
-    void board(){
-
-    }
-
-    static void drawFromDeck(String deck[]) {
-        int deckIndex = 51;
-        for (int i = 0; i < deckIndex; i++) {
-            String drawnCard = deck[i];
-
-
-
-
-        }
-    }
-
-
-}
