--- Assignment1/src/main/java/models/Deal.java
@@ -1,9 +1,0 @@
-package models;
-
-import models.Deck;
-/**
- * Created by Isaac Chan on 1/20/2016.
- */
-public class Deal {
-
-    }
