+++ Assignment1/src/main/java/models/Deck.java
@@ -4,13 +4,14 @@ package models;
  * Created by Isaac Chan on 1/20/2016.
  */
 public class Deck {
-    public static void deck(String[] args) {
+    public static void deck(String[] newDeck) {
         //Hardcoded array of 52 cards
-        String fullDeck[] = {"AH", "2H", "3H", "4H", "5H", "6H", "7H", "8H", "9H", "10H", "JH", "QH", "KH",
+        String[] fullDeck = {"AH", "2H", "3H", "4H", "5H", "6H", "7H", "8H", "9H", "10H", "JH", "QH", "KH",
                 "AD", "2D", "3D", "4D", "5D", "6D", "7D", "8D", "9D", "10D", "JD", "QD", "KD",
                 "AC", "2C", "3C", "4C", "5C", "6C", "7C", "8C", "9C", "10C", "JC", "QC", "KC",
                 "AS", "2S", "3S", "4S", "5S", "6S", "7S", "8S", "9S", "10S", "JS", "QS", "KS",};
 
+        newDeck = fullDeck;
         int N = 52;
 
         for (int i = 0; i < N; i++) {
