+++ Assignment1/src/main/java/models/GameState.java
@@ -1,0 +1,33 @@
+package models;
+import models.Deck;
+import sun.plugin.javascript.navig.Array;
+
+import java.util.ArrayList;
+
+/**
+ * Created by isaac on 1/21/2016.
+ */
+public class GameState {
+    public static void main(String[] args){
+        Deck deck = new Deck();
+
+        static void drawFromDeck(String deck[]) {
+            int deckIndex = 51;
+            for (int i = 0; i < deckIndex; i++) {
+                String drawnCard = deck[i];
+            }
+        }
+
+        String[] col1 = {};
+        String[] col2 = {};
+        String[] col3 = {};
+        String[] col4 = {};
+
+        static void dealCards(String deck[]){
+            col1[0] = drawnCard;
+        }
+    }
+
+
+
+}
